Test mocked serial
To test the serial without connecting to a real device I has to use
"loop://". This device will return errors when trying to read from it
and to prevent the errors a check was added to the digital read method
in the Controller.

To know when to use the loop device I added an environment variable to
the HelperFunctions options.

- add environment to HelperFunctions
- add getter to HelperFunctions
- fix Exception type in DigitalPin constructor
- add check in Connector class to use either testing env or default
- upgrade travis python version to support typing